requirements.txt
elevenlabs
Groq
python-dotenv
agno
streamlit
